requirements.txt
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.41.0
django-bootstrap-static==4.2.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-markdown==0.8.4
gunicorn==20.0.4
lazy-object-proxy==1.4.3
oauthlib==3.1.0
Pillow==7.0.0
